+++ core/Options.lua
@@ -42,7 +42,6 @@ local minimapButtonProxy = setmetatable({}, {
 
 function Options:Initialize()
 	local category, layout = Settings.RegisterVerticalLayoutCategory(addonName)
-	Addon:SetChangelog(HRT.CHANGELOG)
 
 	layout:AddInitializer(CreateSettingsListSectionHeaderInitializer(L["options.general"]))
 
@@ -122,8 +121,7 @@ function Options:Initialize()
 	})
 
 	-- About Section
-	AWL.Settings:AddAboutSection(layout, addonName)
-	Addon:AddChangelogButton(layout)
+	AWL.Settings:AddAboutSection(layout, addonName, HRT.CHANGELOG)
 
 	Settings.RegisterAddOnCategory(category)